website/requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
mysql-connector-python